# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Knowledge Base API",
    "version": "1.0.0",
    "description": "REST API สำหรับจัดการ Knowledge Base - ใช้ x-api-key header สำหรับ authentication"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local" }
  ],
  "security": [{ "ApiKeyAuth": [] }],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
...
  },
  "paths": {
    "/api/v1/kb/search": {
      "get": {
        "summary": "ค้นหา Knowledge และ Flows",
        "operationId": "searchKnowledge",
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string" }, "description": "คำค้นหา" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 3, "maximum": 10 }, "description": "จำนวนผลลัพธ์" }
        ],
        "responses": {
          "200": { "description": "Search results with recommendation" }
        }
      }
    },
    "/api/v1/kb/knowledge": {
      "get": {
        "summary": "รายการ Knowledge Items",
        "operationId": "listKnowledge",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50, "maximum": 100 } },
          { "name": "category_id", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["faq", "info", "procedure"] } },
          { "name": "active_only", "in": "query", "schema": { "type": "boolean", "default": true } }
        ],
        "responses": {
          "200": { "description": "List of knowledge items with pagination" }
        }
      },
      "post": {
        "summary": "สร้าง Knowledge Item ใหม่",
        "operationId": "createKnowledge",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["question", "answer"],
                "properties": {
                  "question": { "type": "string", "maxLength": 500 },
                  "answer": { "type": "string" },
                  "type": { "type": "string", "enum": ["faq", "info", "procedure"], "default": "faq" },
                  "categoryId": { "type": "string", "format": "uuid", "nullable": true },
                  "keywords": { "type": "array", "items": { "type": "string" }, "default": [] },
                  "tags": { "type": "array", "items": { "type": "string" }, "default": [] },
                  "isActive": { "type": "boolean", "default": true }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created knowledge item" }
        }
      }
    },
...
  }
}